Final merge

+++ src/main/java/sopra/formation/repository/IReservationRepository.java
@@ -2,6 +2,6 @@ package sopra.formation.repository;
 
 import sopra.formation.model.Reservation;
 
-public interface IReservationRepository extends IRepository<Reservation, Integer>{
+public interface IReservationRepository extends IRepository<Reservation, Long>{
 
 }
